# tbank/app/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MeowMeow ./\u003c|\u003c3|\u003e\\.",
        "title": "TWish API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/users/{userId}/settings": {
            "get": {
                "description": "Возвращает настройки пользователя по его ID",
                "tags": [
                    "settings"
                ],
                "summary": "Получить настройки пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пользователя",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Settings"
                        }
                    }
                }
            },
            "post": {
                "description": "Сохраняет настройки пользователя по его ID",
                "tags": [
                    "settings"
                ],
                "summary": "Сохранить настройки пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пользователя",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Настройки",
                        "name": "settings",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Settings"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "успешно сохранено",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/{userId}/wishes": {
            "get": {
                "description": "Возвращает все желания пользователя по его ID",
                "tags": [
                    "wishes"
                ],
                "summary": "Получить список желаний пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пользователя",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Wish"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Добавляет новое желание пользователя",
                "tags": [
                    "wishes"
                ],
                "summary": "Добавить желание",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пользователя",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Желание",
                        "name": "wish",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Wish"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "успешно добавлено",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/{userId}/wishes/{wishId}": {
            "delete": {
                "description": "Удаляет желание пользователя по его ID",
                "tags": [
                    "wishes"
                ],
                "summary": "Удалить желание",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пользователя",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID желания",
                        "name": "wishId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "успешно удалено",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/{userId}/wishes/{wishId}/toggle": {
            "post": {
                "description": "Переключает статус StillWant для желания по его ID",
                "tags": [
                    "wishes"
                ],
                "summary": "Переключить статус желания",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пользователя",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID желания",
                        "name": "wishId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "успешно обновлено",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.CooldownRange": {
            "description": "Диапазон охлаждения для настроек.",
            "type": "object",
            "properties": {
                "max": {
                    "description": "Max — максимальное значение\nexample: 10000",
                    "type": "number"
                },
                "min": {
                    "description": "Min — минимальное значение\nexample: 1",
                    "type": "number"
                },
                "period": {
                    "description": "Period — период охлаждения в днях\nexample: 7",
                    "type": "integer"
                }
            }
        },
        "main.Settings": {
            "description": "Настройки пользователя.",
            "type": "object",
            "properties": {
                "cooldowns": {
                    "description": "Cooldowns — диапазоны охлаждения",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.CooldownRange"
                    }
                },
                "excludedProducts": {
                    "description": "ExcludedProducts — исключённые продукты\nexample: \"алкоголь\"",
                    "type": "string"
                },
                "monthlySaving": {
                    "description": "MonthlySaving — ежемесячая сумма\nexample: 300",
                    "type": "number"
                },
                "notificationChannel": {
                    "description": "NotificationChannel — канал уведомлений\nexample: \"email\"",
                    "type": "string"
                },
                "notificationFrequency": {
                    "description": "NotificationFreq — частота уведомлений\nexample: \"еженедельно\"",
                    "type": "string"
                },
                "totalPurchases": {
                    "description": "TotalPurchases — всего покупок\nexample: 25",
                    "type": "integer"
                },
                "totalSpent": {
                    "description": "TotalSpent — потрачено всего\nexample: 1500",
                    "type": "number"
                }
            }
        },
        "main.Wish": {
            "description": "Информация о желании пользователя.",
            "type": "object",
            "properties": {
                "category": {
                    "description": "Category — категория желания\nexample: \"Электроника\"",
                    "type": "string"
                },
                "coolingDays": {
                    "description": "CoolingDays — количество дней на \"остывание\"\nexample: 5",
                    "type": "integer"
                },
                "createdAt": {
                    "description": "CreatedAt — дата и время создания желания\nexample: \"2024-01-01T12:00:00Z\"",
                    "type": "string"
                },
                "id": {
                    "description": "ID — уникальный идентификатор\nexample: \"123e4567-e89b-12d3-a456-426614174000\"",
                    "type": "string"
                },
                "price": {
                    "description": "Price — цена в рублях\nexample: 10 000",
                    "type": "number"
                },
                "recommendedCooling": {
                    "description": "RecommendedCooling — рекомендованное количество дней на \"остывание\"\nexample: 7",
                    "type": "integer"
                },
                "stillWant": {
                    "description": "StillWant — флаг, указывающий, хочет ли пользователь всё ещё это желание\nexample: true",
                    "type": "boolean"
                },
                "title": {
                    "description": "Title — название желания\nexample: \"Новый Ноутбук\"",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}